canary/ping/requirements.txt
kafka-python
python-dotenv
